+++ lib/install/includes/library.sh
@@ -388,7 +388,7 @@ _replaceLineInFile() {
             sed -i "$line_found i $new_string" $file_path            
 
         else
-            echo "ERROR: Target line not found."
+            echo "ERROR: Target line not found for $find_string."
             sleep 2
         fi