+++ lib/install/uninstall.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+clear
+sleep 1
+cp $HOME/.config/ml4w/apps/ML4W_Dotfiles_Uninstaller.AppImage $HOME/.cache
+cd $HOME/.cache
+./ML4W_Dotfiles_Uninstaller.AppImage